LayoutTests/platform/win/accessibility/canvas-fallback-content-expected.txt

Link  Button
Focusable
ARIA button
ARIA link
This test makes sure that focusable elements in canvas fallback content are accessible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


link1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXLink"

button1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"

text1
PASS document.activeElement == element is true
FAIL axElement.role should be AXRole: AXTextField. Was AXRole: AXStaticText.

checkbox1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXCheckBox"

radio1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXRadioButton"

submit1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"

combobox1
PASS document.activeElement == element is true
FAIL axElement.role should be AXRole: AXPopUpButton. Was AXRole: AXComboBox.

focusable1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXGroup"

aria-button1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"

aria-link1
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXLink"

link2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXLink"

button2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"

text2
PASS document.activeElement == element is true
FAIL axElement.role should be AXRole: AXTextField. Was AXRole: AXStaticText.

checkbox2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXCheckBox"

radio2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXRadioButton"

submit2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"

combobox2
PASS document.activeElement == element is true
FAIL axElement.role should be AXRole: AXPopUpButton. Was AXRole: AXComboBox.

focusable2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXGroup"

aria-button2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"

aria-link2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXLink"

focusable1
PASS document.activeElement == element is true
FAIL axElement.role should be AXRole: AXButton. Was AXRole: AXGroup.

focusable2
PASS document.activeElement == element is true
PASS axElement.role is "AXRole: AXButton"

PASS successfullyParsed is true

TEST COMPLETE
